[REVIEW] Share BlazingContext across queries in end-to-end run

+++ tpcx_bb/benchmark_runner.py
@@ -1,45 +1,94 @@
+import glob
+import re
 import os
 import gc
 import time
 
 os.environ["tpcxbb_benchmark_sweep_run"] = "True"
 
-qnums = [str(i).zfill(2) for i in range(1, 31)]
-n_repeats = 3
+N_REPEATS = 1
+
+
+def get_qnum_from_filename(name):
+    m = re.search("[0-9]{2}", name).group()
+    return m
+
+
+dask_qnums = [str(i).zfill(2) for i in range(1, 31)]
+
+# Not all queries are implemented with BSQL
+bsql_query_files = sorted(glob.glob("./queries/q*/t*_sql.py"))
+bsql_qnums = [get_qnum_from_filename(x.split("/")[-1]) for x in bsql_query_files]
 
 
 if __name__ == "__main__":
     from xbb_tools.cluster_startup import attach_to_cluster, import_query_libs
-    from xbb_tools.utils import run_dask_cudf_query, tpcxbb_argparser
+    from xbb_tools.utils import run_query, tpcxbb_argparser
     import importlib
 
     import_query_libs()
 
-    q_func_d = {
+    dask_queries = {
         qnum: importlib.import_module(f"queries.q{qnum}.tpcx_bb_query_{qnum}").main
-        for qnum in qnums
+        for qnum in dask_qnums
+    }
+
+    bsql_queries = {
+        qnum: importlib.import_module(f"queries.q{qnum}.tpcx_bb_query_{qnum}_sql").main
+        for qnum in bsql_qnums
     }
 
     cli_args = tpcxbb_argparser()
-    client = attach_to_cluster(cli_args)
+    
+    include_blazing = cli_args.get("benchmark_runner_include_bsql")
+    client, bc = attach_to_cluster(cli_args, create_blazing_context=include_blazing)
 
     # Preload required libraries for queries on all workers
     client.run(import_query_libs)
 
     base_path = os.getcwd()
-    for qnum, q_func in q_func_d.items():
-        print(qnum)
 
-        qpath = f"{base_path}/queries/q{qnum}/"
-        os.chdir(qpath)
-        if os.path.exists("current_query_num.txt"):
-            os.remove("current_query_num.txt")
-        with open("current_query_num.txt", "w") as fp:
-            fp.write(qnum)
+    # Run Pure Dask Queries
+    if len(dask_qnums) > 0:
+        print("Pure Dask Queries")
+        for qnum, q_func in dask_queries.items():
+            print(qnum)
 
-        for r in range(n_repeats):
-            run_dask_cudf_query(cli_args=cli_args, client=client, query_func=q_func)
-            client.run(gc.collect)
-            client.run_on_scheduler(gc.collect)
-            gc.collect()
-            time.sleep(3)
+            qpath = f"{base_path}/queries/q{qnum}/"
+            os.chdir(qpath)
+            if os.path.exists("current_query_num.txt"):
+                os.remove("current_query_num.txt")
+            with open("current_query_num.txt", "w") as fp:
+                fp.write(qnum)
+
+            for r in range(N_REPEATS):
+                run_query(cli_args=cli_args, client=client, query_func=q_func)
+                client.run(gc.collect)
+                client.run_on_scheduler(gc.collect)
+                gc.collect()
+                time.sleep(3)
+
+    # Run BSQL Queries
+    if include_blazing and len(bsql_qnums) > 0:
+        print("Blazing Queries")
+        for qnum, q_func in bsql_queries.items():
+            print(qnum)
+
+            qpath = f"{base_path}/queries/q{qnum}/"
+            os.chdir(qpath)
+            if os.path.exists("current_query_num.txt"):
+                os.remove("current_query_num.txt")
+            with open("current_query_num.txt", "w") as fp:
+                fp.write(qnum)
+
+            for r in range(N_REPEATS):
+                run_query(
+                    cli_args=cli_args,
+                    client=client,
+                    query_func=q_func,
+                    blazing_context=bc,
+                )
+                client.run(gc.collect)
+                client.run_on_scheduler(gc.collect)
+                gc.collect()
+                time.sleep(3)
